fix: implemented reward union like described in 17.4 of the shelly ledger spec. The reserves are now almost correct with dbsync until epoch 350

+++ src/main/java/org/cardanofoundation/rewards/calculation/EpochCalculation.java
@@ -183,7 +183,7 @@ public class EpochCalculation {
             calculatedReserve = calculatedReserve.add(MAINNET_BOOTSTRAP_ADDRESS_AMOUNT);
         }
 
-        log.debug("Unspendable earned rewards: " + lovelaceToAda(unspendableEarnedRewards.intValue()) + " ADA");
+        log.debug("Unspendable earned rewards: " + lovelaceToAda(unspendableEarnedRewards.longValue()) + " ADA");
         treasuryForCurrentEpoch = add(treasuryForCurrentEpoch, unspendableEarnedRewards);
 
         TreasuryCalculationResult treasuryCalculationResult = TreasuryCalculationResult.builder()
